Delete the transpiler's 85 no-op casts
`unnecessary_cast` was the long tail's biggest entry and the last of the
tree's purely mechanical transpiler residue: casts to the type the
expression already had. Three shapes:

- `1 as usize` / `4096 as isize` on a literal, which is spelled `1_usize`
  and needs no cast at all;
- `p.add(len as usize)` where `len` is already `usize`, and
  `x.offset_from(y) as isize` where `offset_from` returns `isize`;
- a cast to a named type an expression already has --
  `spell_read_aff(..) as *mut afffile_T`, `AUTOMATIC_ENGINE as u32`,
  `kOptFdoFlagSearch as u32`. These are the anonymous C enums the
  transpiler gave concrete types to; the cast is now noise.

Applied from clippy's own suggestion byte spans, so every replacement is
the compiler's, not a regex's. One site had a cast nested inside another
cast's span (`(q.offset_from(p) - 1 as isize) as isize`); the tool grew a
`--skip-nested` flag that keeps the outer edit and leaves the inner one
for the next pass, and the second pass found nothing left.

Three sites had parentheses whose only purpose was binding the deleted
cast; `-D unused-parens` makes those a hard error rather than a warning,
and one import list lost two names that only the deleted casts referred
to. Four hand edits, all in this commit.

clippy 163 -> 78, the first time the tree is under a hundred; unchecked
lines 268,160 -> 268,147. All ten baselined differentials IDENTICAL, all
224 mutation anchors still resolve.

+++ crates/nvim/src/nvim/buffer.rs
@@ -4810,7 +4810,7 @@ pub unsafe extern "C" fn buf_set_changedtick(buf: *mut buf_T, changedtick: varnu
     let changedtick_di: *mut dictitem_T = tv_dict_find(
         (*buf).b_vars,
         c"changedtick".as_ptr(),
-        ::core::mem::size_of::<[::core::ffi::c_char; 12]>().wrapping_sub(1 as usize) as ptrdiff_t,
+        ::core::mem::size_of::<[::core::ffi::c_char; 12]>().wrapping_sub(1_usize) as ptrdiff_t,
     );
     debug_assert!(!changedtick_di.is_null(), "changedtick_di != NULL");
     assert!(
